+++ Documents/Cpp-code/binary.h
@@ -6,14 +6,14 @@
 
 //Tested
 // Find the eccentric anomaly of a binary given its eccentricity e and mean anomaly M
-double eccentricAnomaly(double e, double M);
+long double eccentricAnomaly(long double e, long double M);
 
 //Tested with setupRandomBinary
 //Return the semi-major axis and eccentricity of a binary and whether or not it is bound from the positions and velocities of the stars
-std::tuple<double, double, bool> orbitalElements(std::array<std::array<double, 3>, 4> X, double m1, double m2);
+std::tuple<long double, long double, bool> orbitalElements(std::array<std::array<long double, 3>, 4> X, long double m1, long double m2);
 
 //Tested with orbitalElements
 // Open a binary: find the position and velocity vectors given the semi-major axis and eccentricity
-std::array<std::array<double, 3>, 4> setupRandomBinary(double a, double e, double m1, double m2);
+std::array<std::array<long double, 3>, 4> setupRandomBinary(long double a, long double e, long double m1, long double m2);
 
 #endif
